Make git show ChooseAccountActivity renamed to SignInDelegate.java
