+++ auth/src/main/java/com/firebase/ui/auth/util/smartlock/SignInDelegate.java
@@ -1,154 +1,217 @@
+/*
+ * Copyright 2016 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.firebase.ui.auth.util.smartlock;
 
-import android.app.ProgressDialog;
+import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.content.IntentSender;
+import android.net.ConnectivityManager;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
 import android.text.TextUtils;
 import android.util.Log;
 
-import com.firebase.ui.auth.AuthUI;
+import com.firebase.ui.auth.AuthUI.IdpConfig;
 import com.firebase.ui.auth.BuildConfig;
-import com.firebase.ui.auth.R;
-import com.firebase.ui.auth.provider.IDPProviderParcel;
-import com.firebase.ui.auth.ui.ExtraConstants;
+import com.firebase.ui.auth.ui.ActivityHelper;
 import com.firebase.ui.auth.ui.FlowParameters;
 import com.firebase.ui.auth.ui.TaskFailureLogger;
 import com.firebase.ui.auth.ui.idp.AuthMethodPickerActivity;
-import com.firebase.ui.auth.ui.idp.IDPSignInContainerActivity;
+import com.firebase.ui.auth.ui.idp.IdpSignInContainerActivity;
+import com.firebase.ui.auth.util.CredentialsAPI;
 import com.firebase.ui.auth.util.CredentialsApiHelper;
 import com.firebase.ui.auth.util.EmailFlowUtil;
-import com.firebase.ui.auth.util.FirebaseAuthWrapperFactory;
 import com.firebase.ui.auth.util.PlayServicesHelper;
-import com.google.android.gms.auth.api.Auth;
 import com.google.android.gms.auth.api.credentials.Credential;
-import com.google.android.gms.auth.api.credentials.CredentialRequest;
-import com.google.android.gms.auth.api.credentials.CredentialRequestResult;
+import com.google.android.gms.auth.api.credentials.CredentialsApi;
 import com.google.android.gms.auth.api.credentials.IdentityProviders;
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
-import com.google.android.gms.common.api.CommonStatusCodes;
-import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.android.gms.common.api.Status;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
-import com.google.firebase.FirebaseApp;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.EmailAuthProvider;
 import com.google.firebase.auth.FacebookAuthProvider;
-import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseAuthInvalidUserException;
 import com.google.firebase.auth.GoogleAuthProvider;
-
+import com.google.firebase.auth.TwitterAuthProvider;
 import java.util.List;
 
-import static android.app.Activity.RESULT_CANCELED;
-import static android.app.Activity.RESULT_OK;
+import static com.firebase.ui.auth.ui.ResultCodes.RESULT_NO_NETWORK;
 
-public class SignInDelegate extends SmartLock<CredentialRequestResult> {
-    private static final String TAG = "SignInDelegate";
+/**
+ * Attempts to acquire a credential from Smart Lock for Passwords to sign in
+ * an existing account. If this succeeds, an attempt is made to sign the user in
+ * with this credential. If it does not, the
+ * {@link AuthMethodPickerActivity authentication method picker activity}
+ * is started, unless only email is supported, in which case the
+ * {@link com.firebase.ui.auth.ui.email.SignInNoPasswordActivity email sign-in flow}
+ * is started.
+ */
+public class SignInDelegate extends ActivityBase {
+    private static final String TAG = "ChooseAccountActivity";
+
     private static final int RC_CREDENTIALS_READ = 2;
     private static final int RC_IDP_SIGNIN = 3;
     private static final int RC_AUTH_METHOD_PICKER = 4;
     private static final int RC_EMAIL_FLOW = 5;
     private static final int RC_PLAY_SERVICES = 6;
 
-    private FlowParameters mFlowParams;
-    private ProgressDialog mProgressDialog;
-    private GoogleApiClient mGoogleApiClient;
-    private Credential mCredential;
+    private CredentialsAPI mCredentialsApi;
+    private PlayServicesHelper mPlayServicesHelper;
 
     @Override
-    public void onCreate(Bundle savedInstance) {
+    protected void onCreate(Bundle savedInstance) {
         super.onCreate(savedInstance);
-        setRetainInstance(true); // TODO: 10/16/2016 test low mem devices
 
-        mFlowParams = getArguments().getParcelable(ExtraConstants.EXTRA_FLOW_PARAMS);
+        if (!hasNetworkConnection()) {
+            Log.d(TAG, "No network connection");
+
+            finish(RESULT_NO_NETWORK, new Intent());
+            return;
+        }
 
         // Make Google Play Services available at the correct version, if possible
-        boolean madeAvailable =
-                PlayServicesHelper
-                        .getInstance(getActivity())
-                        .makePlayServicesAvailable(getActivity(), RC_PLAY_SERVICES,
-                                                   new DialogInterface.OnCancelListener() {
-                                                       @Override
-                                                       public void onCancel(DialogInterface dialogInterface) {
-                                                           Log.w(TAG,
-                                                                 "playServices:dialog.onCancel()");
-                                                           finish(RESULT_CANCELED, new Intent());
-                                                       }
-                                                   });
+        mPlayServicesHelper = PlayServicesHelper.getInstance(this);
+        boolean madeAvailable = mPlayServicesHelper
+                .makePlayServicesAvailable(this, RC_PLAY_SERVICES,
+                        new DialogInterface.OnCancelListener() {
+                            @Override
+                            public void onCancel(DialogInterface dialogInterface) {
+                                Log.w(TAG, "playServices:dialog.onCancel()");
+                                finish(RESULT_CANCELED, new Intent());
+                            }
+                        });
 
-        if (!madeAvailable
-                || !PlayServicesHelper.getInstance(getActivity()).isPlayServicesAvailable()
-                || !FirebaseAuthWrapperFactory.getFirebaseAuthWrapper(mFlowParams.appName)
-                .isPlayServicesAvailable(getActivity())) {
+        if (!madeAvailable) {
             Log.w(TAG, "playServices: could not make available.");
             finish(RESULT_CANCELED, new Intent());
             return;
         }
 
-        if (mFlowParams.smartLockEnabled) {
-            showProgress();
-            initGoogleApiClient(null);
-            Auth.CredentialsApi
-                    .request(mGoogleApiClient,
-                             new CredentialRequest.Builder()
-                                     .setPasswordLoginSupported(true)
-                                     .setAccountTypes(IdentityProviders.GOOGLE,
-                                                      IdentityProviders.FACEBOOK,
-                                                      IdentityProviders.TWITTER)
-                                     .build())
-                    .setResultCallback(this);
-        } else {
-            startAuthMethodChoice();
+        mCredentialsApi = new CredentialsAPI(this, new CredentialsAPI.CallbackInterface() {
+            @Override
+            public void onAsyncTaskFinished() {
+                onCredentialsApiConnected(mCredentialsApi, mActivityHelper);
+            }
+        });
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        if (mCredentialsApi != null) {
+            mCredentialsApi.onStart();
         }
     }
 
     @Override
-    public void onResult(@NonNull CredentialRequestResult result) {
-        Status status = result.getStatus();
+    protected void onStop() {
+        super.onStop();
+        if (mCredentialsApi != null) {
+            mCredentialsApi.onStop();
+        }
+    }
 
-        if (status.isSuccess()) {
-            // Auto sign-in success
-            handleCredential(result.getCredential());
-            String email = getEmailFromCredential();
-            String password = getPasswordFromCredential();
+    /**
+     * Check if there is an active or soon-to-be-active network connection.
+     */
+    private boolean hasNetworkConnection() {
+        ConnectivityManager manager =
+                (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
 
-            if (TextUtils.isEmpty(password)) {
-                // log in with id/provider
-                redirectToIdpSignIn(email, getAccountTypeFromCredential());
+        return manager != null
+                && manager.getActiveNetworkInfo() != null
+                && manager.getActiveNetworkInfo().isConnectedOrConnecting();
+    }
+
+    /**
+     * Called when the Credentials API connects.
+     */
+    public void onCredentialsApiConnected(
+            CredentialsAPI credentialsApi,
+            ActivityHelper activityHelper) {
+        String email = credentialsApi.getEmailFromCredential();
+        String password = credentialsApi.getPasswordFromCredential();
+        String accountType = credentialsApi.getAccountTypeFromCredential();
+
+        FlowParameters flowParams = activityHelper.getFlowParams();
+
+        if (flowParams.smartLockEnabled
+                && mPlayServicesHelper.isPlayServicesAvailable()
+                && credentialsApi.isCredentialsAvailable()) {
+
+            // Attempt auto-sign in using SmartLock
+            if (credentialsApi.isAutoSignInAvailable()) {
+                credentialsApi.googleSilentSignIn();
+                if (!TextUtils.isEmpty(password)) {
+                    // Sign in with the email/password retrieved from SmartLock
+                    signInWithEmailAndPassword(activityHelper, email, password);
+                } else {
+                    // log in with id/provider
+                    redirectToIdpSignIn(email, accountType);
+                }
+            } else if (credentialsApi.isSignInResolutionNeeded()) {
+                // resolve credential
+                credentialsApi.resolveSavedEmails(this);
             } else {
-                // Sign in with the email/password retrieved from SmartLock
-                signInWithEmailAndPassword(email, password);
+                startAuthMethodChoice(activityHelper);
             }
-        } else if (status.getStatusCode() == CommonStatusCodes.RESOLUTION_REQUIRED) {
-            hideProgress();
-            // resolve saved emails
-            try {
-                startIntentSenderForResult(status.getResolution().getIntentSender(),
-                                           RC_CREDENTIALS_READ,
-                                           null,
-                                           0,
-                                           0,
-                                           0,
-                                           null);
-            } catch (IntentSender.SendIntentException e) {
-                Log.e(TAG, "Failed to send Credentials intent.", e);
-            }
+        } else {
+            startAuthMethodChoice(activityHelper);
         }
     }
 
-    @Override
-    public void onConnected(@Nullable Bundle bundle) {
+    private void startAuthMethodChoice(ActivityHelper activityHelper) {
+        List<IdpConfig> idpConfigs = activityHelper.getFlowParams().providerInfo;
+
+        // If the only provider is Email, immediately launch the email flow. Otherwise, launch
+        // the auth method picker screen.
+        if (idpConfigs.size() == 1
+                && idpConfigs.get(0).getProviderId().equals(EmailAuthProvider.PROVIDER_ID)) {
+            startActivityForResult(
+                    EmailFlowUtil.createIntent(
+                            this,
+                            activityHelper.getFlowParams()),
+                    RC_EMAIL_FLOW);
+        } else {
+            startActivityForResult(
+                    AuthMethodPickerActivity.createIntent(
+                            this,
+                            activityHelper.getFlowParams()),
+                    RC_AUTH_METHOD_PICKER);
+        }
+    }
+
+    private void logInWithCredential(
+            final String email,
+            final String password,
+            final String accountType) {
+
+        if (email != null
+                && mCredentialsApi.isCredentialsAvailable()
+                && !mCredentialsApi.isSignInResolutionNeeded()) {
+            if (password != null && !password.isEmpty()) {
+                // email/password combination
+                signInWithEmailAndPassword(mActivityHelper, email, password);
+            } else {
+                // identifier/provider combination
+                redirectToIdpSignIn(email, accountType);
+            }
+        }
     }
 
     @Override
@@ -163,28 +226,25 @@ public class SignInDelegate extends SmartLock<CredentialRequestResult> {
                 if (resultCode == RESULT_OK) {
                     // credential selected from SmartLock, log in with that credential
                     Credential credential = data.getParcelableExtra(Credential.EXTRA_KEY);
-                    handleCredential(credential);
-
-                    String email = getEmailFromCredential();
-                    String password = getPasswordFromCredential();
-                    if (email != null) {
-                        if (password == null || password.isEmpty()) {
-                            // identifier/provider combination
-                            redirectToIdpSignIn(email, getAccountTypeFromCredential());
-                        } else {
-                            // email/password combination
-                            signInWithEmailAndPassword(email, password);
-                        }
-                    }
-                } else {
+                    mCredentialsApi.handleCredential(credential);
+                    mCredentialsApi.resolveSignIn();
+                    logInWithCredential(
+                            mCredentialsApi.getEmailFromCredential(),
+                            mCredentialsApi.getPasswordFromCredential(),
+                            mCredentialsApi.getAccountTypeFromCredential()
+                    );
+                } else if (resultCode == RESULT_CANCELED
+                        || resultCode == CredentialsApi.ACTIVITY_RESULT_OTHER_ACCOUNT) {
                     // Smart lock selector cancelled, go to the AuthMethodPicker screen
-                    startAuthMethodChoice();
+                    startAuthMethodChoice(mActivityHelper);
+                } else if (resultCode == RESULT_FIRST_USER) {
+                    // TODO: (serikb) figure out flow
                 }
                 break;
             case RC_IDP_SIGNIN:
             case RC_AUTH_METHOD_PICKER:
             case RC_EMAIL_FLOW:
-                finish(resultCode, new Intent());
+                finish(resultCode, data);
                 break;
             case RC_PLAY_SERVICES:
                 if (resultCode != RESULT_OK) {
@@ -194,92 +254,14 @@ public class SignInDelegate extends SmartLock<CredentialRequestResult> {
         }
     }
 
-    private String getEmailFromCredential() {
-        if (mCredential == null) {
-            return null;
-        }
-        return mCredential.getId();
-    }
-
-    private String getAccountTypeFromCredential() {
-        if (mCredential == null) {
-            return null;
-        }
-        return mCredential.getAccountType();
-    }
-
-    private String getPasswordFromCredential() {
-        if (mCredential == null) {
-            return null;
-        }
-        return mCredential.getPassword();
-    }
-
-    private void handleCredential(Credential credential) {
-        mCredential = credential;
-
-        if (IdentityProviders.GOOGLE.equals(credential.getAccountType())) {
-            // Google account, rebuild GoogleApiClient to set account name and then try
-            initGoogleApiClient(credential.getId());
-            // Try silent sign-in with Google Sign In API
-            Auth.GoogleSignInApi.silentSignIn(mGoogleApiClient);
-        } else {
-            // Email/password account
-            String status = String.format("Signed in as %s", credential.getId());
-            Log.d(TAG, status);
-        }
-    }
-
-    private void initGoogleApiClient(String accountName) {
-        GoogleSignInOptions.Builder gsoBuilder = new GoogleSignInOptions
-                .Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestEmail();
-
-        if (accountName != null) {
-            gsoBuilder.setAccountName(accountName);
-        }
-
-        if (mGoogleApiClient != null) {
-            mGoogleApiClient.disconnect();
-        }
-
-        mGoogleApiClient = new GoogleApiClient.Builder(getContext())
-                .addConnectionCallbacks(this)
-                .addApi(Auth.CREDENTIALS_API)
-                .addApi(Auth.GOOGLE_SIGN_IN_API, gsoBuilder.build())
-                .build();
-        mGoogleApiClient.connect();
-    }
-
-    private void startAuthMethodChoice() {
-        List<IDPProviderParcel> providers = mFlowParams.providerInfo;
-
-        // If the only provider is Email, immediately launch the email flow. Otherwise, launch
-        // the auth method picker screen.
-        if (providers.size() == 1
-                && providers.get(0).getProviderType().equals(EmailAuthProvider.PROVIDER_ID)) {
-            startActivityForResult(
-                    EmailFlowUtil.createIntent(
-                            getContext(),
-                            mFlowParams),
-                    RC_EMAIL_FLOW);
-        } else {
-            startActivityForResult(
-                    AuthMethodPickerActivity.createIntent(
-                            getContext(),
-                            mFlowParams),
-                    RC_AUTH_METHOD_PICKER);
-        }
-    }
-
     /**
      * Begin sign in process with email and password from a SmartLock credential.
-     * On success, finish with {@code RESULT_OK}.
+     * On success, finish with {@link #RESULT_OK}.
      * On failure, delete the credential from SmartLock (if applicable) and then launch the
      * auth method picker flow.
      */
-    private void signInWithEmailAndPassword(String email, String password) {
-        FirebaseAuth.getInstance(FirebaseApp.getInstance(mFlowParams.appName))
+    private void signInWithEmailAndPassword(ActivityHelper helper, String email, String password) {
+        helper.getFirebaseAuth()
                 .signInWithEmailAndPassword(email, password)
                 .addOnFailureListener(new TaskFailureLogger(
                         TAG, "Error signing in with email and password"))
@@ -298,36 +280,10 @@ public class SignInDelegate extends SmartLock<CredentialRequestResult> {
                             // before continuing.
                             deleteCredentialAndRedirect();
                         } else {
-                            startAuthMethodChoice();
+                            startAuthMethodChoice(mActivityHelper);
                         }
                     }
                 });
-    }
-
-    private void redirectToIdpSignIn(String email, String accountType) {
-        Intent nextIntent;
-        switch (accountType) {
-            case IdentityProviders.GOOGLE:
-                nextIntent = IDPSignInContainerActivity.createIntent(
-                        getContext(),
-                        mFlowParams,
-                        GoogleAuthProvider.PROVIDER_ID,
-                        email);
-                break;
-            case IdentityProviders.FACEBOOK:
-                nextIntent = IDPSignInContainerActivity.createIntent(
-                        getContext(),
-                        mFlowParams,
-                        FacebookAuthProvider.PROVIDER_ID,
-                        email);
-                break;
-            default:
-                Log.w(TAG, "unknown provider: " + accountType);
-                nextIntent = AuthMethodPickerActivity.createIntent(
-                        getContext(),
-                        mFlowParams);
-        }
-        this.startActivityForResult(nextIntent, RC_IDP_SIGNIN);
     }
 
     /**
@@ -335,74 +291,61 @@ public class SignInDelegate extends SmartLock<CredentialRequestResult> {
      * auth method choice flow.
      */
     private void deleteCredentialAndRedirect() {
-        if (mCredential == null) {
+        if (mCredentialsApi.getCredential() == null) {
             Log.w(TAG, "deleteCredentialAndRedirect: null credential");
-            startAuthMethodChoice();
+            startAuthMethodChoice(mActivityHelper);
             return;
         }
 
-        CredentialsApiHelper credentialsApiHelper = CredentialsApiHelper.getInstance(getActivity());
-        credentialsApiHelper.delete(mCredential)
-                .addOnCompleteListener(getActivity(), new OnCompleteListener<Status>() {
+        CredentialsApiHelper credentialsApiHelper = CredentialsApiHelper.getInstance(this);
+        credentialsApiHelper.delete(mCredentialsApi.getCredential())
+                .addOnCompleteListener(this, new OnCompleteListener<Status>() {
                     @Override
                     public void onComplete(@NonNull Task<Status> task) {
                         if (!task.isSuccessful()) {
                             Log.w(TAG, "deleteCredential:failure", task.getException());
                         }
-                        startAuthMethodChoice();
+                        startAuthMethodChoice(mActivityHelper);
                     }
                 });
     }
 
-    private void showProgress() {
-        if (mProgressDialog == null || !mProgressDialog.isShowing()) {
-            mProgressDialog = new ProgressDialog(getActivity());
-            mProgressDialog.setIndeterminate(true);
-            mProgressDialog.setMessage(getString(R.string.progress_dialog_loading));
+    protected void redirectToIdpSignIn(String email, String accountType) {
+        Intent nextIntent;
+        switch (accountType) {
+            case IdentityProviders.GOOGLE:
+                nextIntent = IdpSignInContainerActivity.createIntent(
+                        this,
+                        mActivityHelper.getFlowParams(),
+                        GoogleAuthProvider.PROVIDER_ID,
+                        email);
+                break;
+            case IdentityProviders.FACEBOOK:
+                nextIntent = IdpSignInContainerActivity.createIntent(
+                        this,
+                        mActivityHelper.getFlowParams(),
+                        FacebookAuthProvider.PROVIDER_ID,
+                        email);
+                break;
+            case IdentityProviders.TWITTER:
+                nextIntent = IdpSignInContainerActivity.createIntent(
+                        this,
+                        mActivityHelper.getFlowParams(),
+                        TwitterAuthProvider.PROVIDER_ID,
+                        email);
+                break;
+            default:
+                Log.w(TAG, "unknown provider: " + accountType);
+                nextIntent = AuthMethodPickerActivity.createIntent(
+                        this,
+                        mActivityHelper.getFlowParams());
         }
-        mProgressDialog.show();
+        this.startActivityForResult(nextIntent, RC_IDP_SIGNIN);
     }
 
-    private void hideProgress() {
-        if (mProgressDialog != null && mProgressDialog.isShowing()) {
-            mProgressDialog.dismiss();
-        }
-    }
-
-    private void finish(int resultCode, Intent data) {
-        if (mGoogleApiClient != null) {
-            mGoogleApiClient.disconnect();
-        }
-        hideProgress();
-
-        try {
-            ((AuthUI.AuthUIResult) getActivity()).onResult(resultCode, data);
-        } catch (ClassCastException e) {
-            Log.e(TAG, getActivity().toString() + " AuthUI.AuthUIResult");
-        }
-
-        getActivity().getSupportFragmentManager().beginTransaction().remove(this).commit();
-    }
-
-    public static SignInDelegate newInstance(FragmentActivity activity, FlowParameters parameters) {
-        SignInDelegate result;
-
-        FragmentManager fm = activity.getSupportFragmentManager();
-        FragmentTransaction ft = fm.beginTransaction();
-
-        Fragment fragment = fm.findFragmentByTag(TAG);
-        if (fragment == null || !(fragment instanceof SignInDelegate)) {
-            result = new SignInDelegate();
-
-            Bundle bundle = new Bundle();
-            bundle.putParcelable(ExtraConstants.EXTRA_FLOW_PARAMS, parameters);
-            result.setArguments(bundle);
-
-            ft.add(result, TAG).disallowAddToBackStack().commit();
-        } else {
-            result = (SignInDelegate) fragment;
-        }
-
-        return result;
+    public static Intent createIntent(
+            Context context,
+            FlowParameters flowParams) {
+        return ActivityHelper.createBaseIntent(context, ChooseAccountActivity.class, flowParams);
     }
 }
